finished review and start rating
