--- front_end/src/reducers/reviewReducer.js
@@ -1,21 +1,0 @@
-// import { PUBLISH_REVIEW } from "../actions/types";
-
-// const initialState = {
-//   review: {
-//     starReview: 0,
-//     comment: ""
-//   }
-// };
-
-// // ...state = current state
-// export default function(state = initialState, action) {
-//   switch (action.type) {
-//     case PUBLISH_REVIEW:
-//       return {
-//         ...state,
-//         review: action.payload
-//       };
-//     default:
-//       return state;
-//   }
-// }
